feat: Add Influx system to Config
Added the Influx system to the Config struct to manage InfluxDB configurations.
Also added functions to set up and build InfluxDB connections.
